requirements.txt
capstone
lief
pyelftools
click
numpy
scipy
binwalk
rich
modules
python-magic-bin